requirements.txt
dj-database-url==0.4.1
Django==1.11
djangorestframework==3.6.2
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
MySQL-python==1.2.5
Flask==0.11
Jinja2==2.8
gunicorn==19.6.0